moving implementation of MakeBoxes to mathics.eval.makeboxes. Splitting this module in package submodules.
