+++ mathics/eval/makeboxes/formatvalues.py
@@ -8,7 +8,7 @@ formatting rules.
 
 from typing import Any, Callable, Dict, List, Optional, Type
 
-from mathics.core.atoms import Complex, Integer, Rational, Real, String, SymbolI
+from mathics.core.atoms import Complex, Integer, Rational, String, SymbolI
 from mathics.core.convert.expression import to_expression_with_specialization
 from mathics.core.element import BaseElement, BoxElementMixin, EvalMixin
 from mathics.core.evaluation import Evaluation
@@ -23,7 +23,6 @@ from mathics.core.symbols import (
     SymbolGraphics3D,
     SymbolHoldForm,
     SymbolList,
-    SymbolMakeBoxes,
     SymbolNumberForm,
     SymbolPlus,
     SymbolPostfix,
@@ -31,13 +30,7 @@ from mathics.core.symbols import (
     SymbolRepeatedNull,
     SymbolTimes,
 )
-from mathics.core.systemsymbols import (
-    SymbolComplex,
-    SymbolMinus,
-    SymbolRational,
-    SymbolRowBox,
-    SymbolStandardForm,
-)
+from mathics.core.systemsymbols import SymbolComplex, SymbolMinus, SymbolRational
 
 # These Strings are used in Boxing output
 StringElipsis = String("...")
@@ -45,182 +38,19 @@ StringLParen = String("(")
 StringRParen = String(")")
 StringRepeated = String("..")
 
-builtins_precedence: Dict[Symbol, int] = {}
 
-element_formatters: Dict[
+# do_format_*
+
+_element_formatters: Dict[
     Type[BaseElement],
     Callable[[BaseElement, Evaluation, Symbol], Optional[BaseElement]],
 ] = {}
 
 
-# this temporarily replaces the _BoxedString class
-def _boxed_string(string: str, **options):
-    from mathics.builtin.box.layout import StyleBox
-
-    return StyleBox(String(string), **options)
-
-
-def compare_precedence(
-    element: BaseElement, precedence: Optional[int] = None
-) -> Optional[int]:
-    """
-    compare the precedence of the element regarding a precedence value.
-    If both precedences are equal, return 0. If precedence of the
-    first element is higher, return 1, otherwise -1.
-    If precedences cannot be compared, return None.
-    """
-    while element.has_form("HoldForm", 1):
-        element = element.elements[0]
-
-    if precedence is None:
-        return None
-    if element.has_form(("Infix", "Prefix", "Postfix"), 3, None):
-        element_prec = element.elements[2].value
-    elif element.has_form("PrecedenceForm", 2):
-        element_prec = element.elements[1].value
-    # For negative values, ensure that the element_precedence is at least the precedence. (Fixes #332)
-    elif isinstance(element, (Integer, Real)) and element.value < 0:
-        element_prec = precedence
-    else:
-        element_prec = builtins_precedence.get(element.get_head_name())
-
-    if element_prec is None:
-        return None
-    return 0 if element_prec == precedence else (1 if element_prec > precedence else -1)
-
-
-# 640 = sys.int_info.str_digits_check_threshold.
-# Someday when 3.11 is the minimum version of Python supported,
-# we can replace the magic value 640 below with sys.int.str_digits_check_threshold.
-def int_to_string_shorter_repr(value: int, form: Symbol, max_digits=640):
-    """Convert value to a String, restricted to max_digits characters.
-
-    if value has an n-digit decimal representation,
-      value = d_1 *10^{n-1} d_2 * 10^{n-2} + d_3 10^{n-3} + ..... +
-              d_{n-2}*100 +d_{n-1}*10 + d_{n}
-    is represented as the string
-
-    "d_1d_2d_3...d_{k}<<n-2k>>d_{n-k-1}...d_{n-2}d_{n-1}d_{n}"
-
-    where n-2k digits are replaced by a placeholder.
-    """
-    if max_digits == 0:
-        return String(str(value))
-
-    # Normalize to positive quantities
-    is_negative = value < 0
-    if is_negative:
-        value = -value
-        max_digits = max_digits - 1
-
-    # Estimate the number of decimal digits
-    num_digits = int(value.bit_length() * 0.3)
-
-    # If the estimated number is below the threshold,
-    # return it as it is.
-    if num_digits <= max_digits:
-        if is_negative:
-            return String("-" + str(value))
-        return String(str(value))
-
-    # estimate the size of the placeholder
-    size_placeholder = len(str(num_digits)) + 6
-    # Estimate the number of available decimal places
-    avaliable_digits = max(max_digits - size_placeholder, 0)
-    # how many most significative digits include
-    len_msd = (avaliable_digits + 1) // 2
-    # how many least significative digits to include:
-    len_lsd = avaliable_digits - len_msd
-    # Compute the msd.
-    msd = str(value // 10 ** (num_digits - len_msd))
-    if msd == "0":
-        msd = ""
-
-    # If msd has more digits than the expected, it means that
-    # num_digits was wrong.
-    extra_msd_digits = len(msd) - len_msd
-    if extra_msd_digits > 0:
-        # Remove the extra digit and fix the real
-        # number of digits.
-        msd = msd[:len_msd]
-        num_digits = num_digits + 1
-
-    lsd = ""
-    if len_lsd > 0:
-        lsd = str(value % 10 ** (len_lsd))
-        # complete decimal positions in the lsd:
-        lsd = (len_lsd - len(lsd)) * "0" + lsd
-
-    # Now, compute the true number of hiding
-    # decimal places, and built the placeholder
-    remaining = num_digits - len_lsd - len_msd
-    placeholder = f" <<{remaining}>> "
-    # Check if the shorten string is actually
-    # shorter than the full string representation:
-    if len(placeholder) < remaining:
-        value_str = f"{msd}{placeholder}{lsd}"
-    else:
-        value_str = str(value)
-
-    if is_negative:
-        value_str = "-" + value_str
-    return String(value_str)
-
-
-def eval_fullform_makeboxes(
-    self, expr, evaluation: Evaluation, form=SymbolStandardForm
-) -> Optional[BaseElement]:
-    """
-    This function takes the definitions provided by the evaluation
-    object, and produces a boxed form for expr.
-
-    Basically: MakeBoxes[expr // FullForm]
-    """
-    # This is going to be reimplemented.
-    expr = Expression(SymbolFullForm, expr)
-    return Expression(SymbolMakeBoxes, expr, form).evaluate(evaluation)
-
-
-def eval_makeboxes(
-    expr, evaluation: Evaluation, form=SymbolStandardForm
-) -> Optional[BaseElement]:
-    """
-    This function takes the definitions provided by the evaluation
-    object, and produces a boxed fullform for expr.
-
-    Basically: MakeBoxes[expr // form]
-    """
-    # This is going to be reimplemented.
-    return Expression(SymbolMakeBoxes, expr, form).evaluate(evaluation)
-
-
-def format_element(
-    element: BaseElement, evaluation: Evaluation, form: Symbol, **kwargs
-) -> Optional[BaseElement]:
-    """
-    Applies formats associated to the expression, and then calls Makeboxes
-    """
-    evaluation.is_boxing = True
-    expr = do_format(element, evaluation, form)
-    if expr is None:
-        return None
-    result = Expression(SymbolMakeBoxes, expr, form)
-    result_box = result.evaluate(evaluation)
-    if isinstance(result_box, String):
-        return result_box
-    if isinstance(result_box, BoxElementMixin):
-        return result_box
-    else:
-        return format_element(element, evaluation, SymbolFullForm, **kwargs)
-
-
-# do_format_*
-
-
 def do_format(
     element: BaseElement, evaluation: Evaluation, form: Symbol
 ) -> Optional[BaseElement]:
-    do_format_method = element_formatters.get(type(element), do_format_element)
+    do_format_method = _element_formatters.get(type(element), do_format_element)
     return do_format_method(element, evaluation, form)
 
 
@@ -299,7 +129,7 @@ def do_format_element(
 
         formatted = format_expr(expr) if isinstance(expr, EvalMixin) else None
         if formatted is not None:
-            do_format_fn = element_formatters.get(type(formatted), do_format_element)
+            do_format_fn = _element_formatters.get(type(formatted), do_format_element)
             result = do_format_fn(formatted, evaluation, form)
             if include_form and result is not None:
                 result = Expression(form, result)
@@ -317,7 +147,7 @@ def do_format_element(
             # just return it as it is.
             if len(expr.get_elements()) != 1:
                 return expr
-            do_format_fn = element_formatters.get(type(element), do_format_element)
+            do_format_fn = _element_formatters.get(type(element), do_format_element)
             result = do_format_fn(expr, evaluation, form)
             if isinstance(result, Expression):
                 expr = result
@@ -329,14 +159,14 @@ def do_format_element(
         ):
             new_elements = tuple(
                 (
-                    element_formatters.get(type(element), do_format_element)(
+                    _element_formatters.get(type(element), do_format_element)(
                         element, evaluation, form
                     )
                     for element in expr.elements
                 )
             )
             expr_head = expr.head
-            do_format = element_formatters.get(type(expr_head), do_format_element)
+            do_format = _element_formatters.get(type(expr_head), do_format_element)
             head = do_format(expr_head, evaluation, form)
             expr = to_expression_with_specialization(head, *new_elements)
 
@@ -371,7 +201,8 @@ def do_format_rational(
         if minus:
             result = Expression(SymbolMinus, result)
         result = Expression(SymbolHoldForm, result)
-        return do_format_expression(result, evaluation, form)
+        result = do_format_expression(result, evaluation, form)
+        return result
 
 
 def do_format_complex(
@@ -425,29 +256,6 @@ def do_format_expression(
     return expr
 
 
-def parenthesize(
-    precedence: Optional[int],
-    element: Expression,
-    element_boxes,
-    when_equal: bool,
-) -> Expression:
-    """
-    "Determines if ``element_boxes`` needs to be surrounded with parenthesis.
-    This is done based on ``precedence`` and the computed preceence of
-    ``element``.  The adjusted ListExpression is returned.
-
-    If when_equal is True, parentheses will be added if the two
-    precedence values are equal.
-    """
-    cmp = compare_precedence(element, precedence)
-    if cmp is not None and (cmp == -1 or cmp == 0 and when_equal):
-        return Expression(
-            SymbolRowBox,
-            ListExpression(String("("), element_boxes, String(")")),
-        )
-    return element_boxes
-
-
-element_formatters[Rational] = do_format_rational
-element_formatters[Complex] = do_format_complex
-element_formatters[Expression] = do_format_expression
+_element_formatters[Rational] = do_format_rational
+_element_formatters[Complex] = do_format_complex
+_element_formatters[Expression] = do_format_expression
